stop label update a bit after animation
